requirements.txt
streamlit
langchain
langchain-community
google-genai
faiss-cpu
pypdf
python-dotenv
